+++ src/doitoml/errors.py
@@ -26,6 +26,16 @@ class ConfigError(DoitomlError):
     """An error related to configuration."""
 
 
+class UnsafePathError(ConfigError):
+
+    """An error related to unsafe paths."""
+
+
+class SkipError(ConfigError):
+
+    """An error related to an ambiguous skip."""
+
+
 class NoConfigError(ConfigError):
 
     """An error when no configuration at all is found."""