+++ CHANGELOG.md
@@ -1,8 +1,33 @@
 # Changelog
 
+## 0.2.0 (unreleased)
+
+### Data Model
+
+- expands `skip` to allow custom skippers like `any`, `all`, `not`, `exists`, `platform`
+
+### Docs
+
+- adds interactive demo and playground, powered by JupyterLite
+
+### API
+
+- adds `safe_paths` to the Python API
+  - arbitrary paths in _shell_ actions are uncheckable, however
+
+### Packaging
+
+- adds `doitoml[test]`
+
 ## 0.1.1
 
-- fix schema import
+### Bugs
+
+- fixes `jsonschema` dependency
+
+### Packaging
+
+- adds `doitoml[all]`
 
 ## 0.1.0